refactor(sandbox): extract gateway recovery preload self-heal builder
